refactor : Grade 도메인 패키키 변경 및 3tier 기본 클래스 생성
