+++ src/main/java/com/jasik/momsnaggingapi/domain/grade/controller/GradeController.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.grade.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class GradeController {
+}
